naming and comments
