+++ tests/test_server_rename.py
@@ -8,13 +8,13 @@ from lsprotocol import types as lsp
 from xonsh_lsp import server as server_module
 
 
-def _make_workspace_edit(uri: str, ranges: list[lsp.Range], new_name: str) -> lsp.WorkspaceEdit:
+def _build_workspace_edit(uri: str, ranges: list[lsp.Range], new_name: str) -> lsp.WorkspaceEdit:
     return lsp.WorkspaceEdit(
         changes={uri: [lsp.TextEdit(range=r, new_text=new_name) for r in ranges]}
     )
 
 
-class TestRename:
+class TestServerRename:
     """Test textDocument/rename support."""
 
     @pytest.fixture
@@ -28,7 +28,7 @@ class TestRename:
     def patched_server(self, monkeypatch, mock_doc):
         """Patch the module-level server with a mock backend.rename."""
         uri = "file:///test/file.xsh"
-        backend_edit = _make_workspace_edit(
+        backend_edit = _build_workspace_edit(
             uri,
             [
                 lsp.Range(
@@ -127,27 +127,3 @@ class TestRename:
         result = await server_module.rename(params)
 
         assert result is None
-
-
-class TestJediBackendRename:
-    """End-to-end check that JediBackend produces a real WorkspaceEdit."""
-
-    @pytest.mark.asyncio
-    async def test_jedi_rename_produces_expected_edits(self):
-        from xonsh_lsp.jedi_backend import JEDI_AVAILABLE, JediBackend
-
-        if not JEDI_AVAILABLE:
-            pytest.skip("jedi not installed")
-
-        backend = JediBackend()
-        src = "value = 1\nprint(value)\nvalue = value + 1\n"
-        edit = await backend.rename(src, 0, 2, "renamed", path=None)
-
-        assert edit is not None
-        assert edit.changes is not None
-        edits = next(iter(edit.changes.values()))
-        ranges = sorted(
-            (e.range.start.line, e.range.start.character) for e in edits
-        )
-        assert ranges == [(0, 0), (1, 6), (2, 0), (2, 8)]
-        assert all(e.new_text == "renamed" for e in edits)
